Start of part 2
I am working on splitting the lines into groups of threes. I know theres going to be an easier solution then what I am working on, but I do not want to cheat. My solution is to put each line into an array and plus a counter. On the third count, find the duplicate letter in the array and add that letter to @letters array and find the piroirty of the letter.