Day 4, part 1 solution
I was able to get rid of the bug, break was causing the entire loop to end too early and not test the rest of input.md. I then added a elsif to the second ranges test to make sure no single line would be counted twice toward @redundant_assignment_pairs.